Publish fewer metrics
Changes the reporting interval to 15 seconds, which drops our bill per metric
from $0.25/source/month to $0.20/source/month.

Meters publish a counter and also publish exponentially weighted moving
averages as Gauges, which are overkill and also increase the number of metrics
we're sending to Librato. Instead just send a Counter; I think I had it like
this initially because I was troubleshooting the metric displays as part of
a separate problem.